Update Go article with API key for OpenWeatherMap
This fixes the Go article by peterbourgon to have info and code about
using the OpenWeatherMap API key that is now required. I noticed this by
going through the article, and the adjustments are fairly minor.

The reason for this change is that starting on October 9, 2015,
OpenWeatherMap started to require an API key (a.k.a. APPID) to make
requests. See http://openweathermap.org/faq for more info on this
requirement and http://openweathermap.org/appid on using the APPID.

The new code in the article requires the person going through it to
insert their API key where it says "YOUR_API_KEY". The requests to
OpenWeatherMap will then work as expected.